Break up WorkspaceSnapshotGraph & Migrator into versioned module namespacing
Since we're planning on moving to a graph V3 in the near future, this
helps create a structure to keep things more organized than having every
version ever in the same file.

Author:    Jacob Helwig <jacob@technosorcery.net>
